refactor: css를 별도의 파일로 분리 (#435)
