+++ frontend/src/layouts/SidebarLayout/components/Sidebar/Sidebar.tsx
@@ -1,5 +1,4 @@
 import { useState, useEffect } from 'react';
-import styled from '@emotion/styled';
 import { useNavigate, useParams } from 'react-router-dom';
 import Logo from '../../../../assets/logo.svg';
 
@@ -15,6 +14,7 @@ import SidebarFeatureMenu from '../SidebarFeatureMenu/SidebarFeatureMenu';
 import SidebarInvitationMenu from '../SidebarInvitationMenu/SidebarInvitationMenu';
 import SidebarSlackMenu from '../SidebarSlackMenu/SidebarSlackMenu';
 import SidebarLogoutMenu from '../SidebarLogoutMenu/SidebarLogoutMenu';
+import { StyledContainer, StyledLogo, StyledBottomMenu } from './Sidebar.styles';
 
 function Sidebar() {
   const [isLoading, setIsLoading] = useState(true);
@@ -86,35 +86,5 @@ function Sidebar() {
     </>
   );
 }
-
-const StyledContainer = styled.div(
-  ({ theme }) => `
-  position: sticky;
-  top: 0;
-  width: 36.4rem;
-  height: 100vh;
-  z-index: 1; 
-  background: ${theme.colors.WHITE_100};
-  padding-left: 4rem;
-  gap: 2rem;
-`
-);
-
-const StyledLogo = styled.img`
-  display: block;
-  margin: 2rem auto;
-  width: 16rem;
-  aspect-ratio: 16 / 9;
-  cursor: pointer;
-  padding-right: 4rem;
-`;
-
-const StyledBottomMenu = styled.div`
-  position: absolute;
-  bottom: 4rem;
-  display: flex;
-  flex-direction: column;
-  gap: 2rem;
-`;
 
 export default Sidebar;
