+++ frontend/src/layouts/SidebarLayout/components/SidebarFeatureMenu/SidebarFeatureMenu.styles.tsx
@@ -1,0 +1,87 @@
+import styled from '@emotion/styled';
+
+const StyledMenuHeader = styled.div`
+  font-size: 1.6rem;
+  text-align: left;
+  margin-bottom: 2rem;
+`;
+
+const StyledFeatureContainer = styled.div`
+  margin: 2.8rem 0;
+`;
+
+const StyledPollIcon = styled.img`
+  width: 2.4rem;
+`;
+
+const StyledAppointmentIcon = styled.img`
+  width: 2.4rem;
+`;
+
+const StyledFeatureTitle = styled.div`
+  font-size: 1.6rem;
+`;
+
+const StyledPollMenu = styled.div<{
+  isActive: boolean;
+}>(
+  ({ isActive, theme }) => `
+    display: flex;
+    gap: 2rem;
+    align-items: center;
+    cursor: pointer;
+    padding: 2rem;
+
+  ${
+    isActive &&
+    `
+      background: ${theme.colors.GRAY_100}; 
+      border-top-left-radius: 4rem; 
+      border-bottom-left-radius: 4rem;
+    `
+  };
+
+  &:hover {
+    background: ${!isActive && theme.colors.TRANSPARENT_GRAY_100_80};
+    border-top-left-radius: 4rem; 
+    border-bottom-left-radius: 4rem;
+  } 
+`
+);
+
+const StyledAppointmentMenu = styled.div<{
+  isActive: boolean;
+}>(
+  ({ isActive, theme }) => `
+    display: flex;
+    gap: 2rem;
+    align-items: center;
+    cursor: pointer;
+    padding: 2rem;
+  
+  ${
+    isActive &&
+    `
+      background: ${theme.colors.GRAY_100}; 
+      border-top-left-radius: 4rem; 
+      border-bottom-left-radius: 4rem;
+  `
+  };
+
+  &:hover {
+    background: ${!isActive && theme.colors.TRANSPARENT_GRAY_100_80};
+    border-top-left-radius: 4rem; 
+    border-bottom-left-radius: 4rem;
+  } 
+`
+);
+
+export {
+  StyledMenuHeader,
+  StyledFeatureContainer,
+  StyledPollIcon,
+  StyledAppointmentIcon,
+  StyledFeatureTitle,
+  StyledPollMenu,
+  StyledAppointmentMenu
+};
