+++ frontend/src/layouts/SidebarLayout/components/SidebarInvitationMenu/SidebarInvitationMenu.styles.tsx
@@ -1,0 +1,19 @@
+import styled from '@emotion/styled';
+
+const StyledInvitationLink = styled.button`
+  display: flex;
+  align-items: center;
+  gap: 1.2rem;
+  font-size: 1.6rem;
+  padding: 0.4rem;
+`;
+
+const StyledInvitationText = styled.p`
+  font-size: 1.6rem;
+`;
+
+const StyledPlusIcon = styled.img`
+  width: 2.4rem;
+`;
+
+export { StyledInvitationLink, StyledInvitationText, StyledPlusIcon };
