+++ frontend/src/layouts/SidebarLayout/components/SidebarLogoutMenu/SidebarLogoutMenu.tsx
@@ -1,7 +1,7 @@
-import styled from '@emotion/styled';
 import { useNavigate } from 'react-router-dom';
 import Leave from '../../../../assets/leave.svg';
 import { removeLocalStorageItem } from '../../../../utils/storage';
+import { StyledContainer, StyledLeaveImage, StyledText } from './SidebarLogoutMenu.styles';
 
 function SidebarLogoutMenu() {
   const navigate = useNavigate();
@@ -20,25 +20,5 @@ function SidebarLogoutMenu() {
     </StyledContainer>
   );
 }
-
-const StyledContainer = styled.button`
-  display: flex;
-  align-items: center;
-  cursor: pointer;
-  gap: 1.2rem;
-  font-size: 1.6rem;
-`;
-
-const StyledLeaveImage = styled.img`
-  width: 2rem;
-  filter: invert(65%) sepia(7%) saturate(12%) hue-rotate(1deg) brightness(92%) contrast(84%);
-`;
-
-const StyledText = styled.p(
-  ({ theme }) => `
-  font-size: 1.6rem; 
-  color: ${theme.colors.GRAY_400}
-`
-);
 
 export default SidebarLogoutMenu;
