+++ frontend/src/layouts/SidebarLayout/components/SidebarMembersProfileMenu/SidebarMembersProfileMenu.styles.tsx
@@ -1,0 +1,22 @@
+import styled from '@emotion/styled';
+
+const StyledMemberListContainer = styled.div`
+  margin: 2.8rem 0;
+`;
+
+const StyledMenuHeader = styled.div`
+  font-size: 1.6rem;
+  text-align: left;
+  margin-bottom: 2rem;
+`;
+
+const StyledName = styled.div`
+  font-size: 1.6rem;
+`;
+
+const StyledContainer = styled.div`
+  overflow-y: auto;
+  max-height: 36rem;
+`;
+
+export { StyledMemberListContainer, StyledMenuHeader, StyledName, StyledContainer };
