+++ frontend/src/layouts/SidebarLayout/components/SidebarMembersProfileMenu/SidebarMembersProfileMenu.tsx
@@ -1,10 +1,15 @@
-import styled from '@emotion/styled';
 import { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { getGroupMembers } from '../../../../api/group';
 import { GroupInterface, MemberInterface } from '../../../../types/group';
 import Avatar from '../../../../components/Avatar/Avatar';
 import FlexContainer from '../../../../components/FlexContainer/FlexContainer';
+import {
+  StyledMemberListContainer,
+  StyledMenuHeader,
+  StyledName,
+  StyledContainer
+} from './SidebarMembersProfileMenu.styles';
 
 interface Props {
   groupCode: GroupInterface['code'];
@@ -60,24 +65,5 @@ function SidebarMembersProfileMenu({ groupCode }: Props) {
     </StyledMemberListContainer>
   );
 }
-
-const StyledMemberListContainer = styled.div`
-  margin: 2.8rem 0;
-`;
-
-const StyledMenuHeader = styled.div`
-  font-size: 1.6rem;
-  text-align: left;
-  margin-bottom: 2rem;
-`;
-
-const StyledName = styled.div`
-  font-size: 1.6rem;
-`;
-
-const StyledContainer = styled.div`
-  overflow-y: auto;
-  max-height: 36rem;
-`;
 
 export default SidebarMembersProfileMenu;
