+++ frontend/src/layouts/SidebarLayout/components/SidebarSlackMenu/SidebarSlackMenu.styles.tsx
@@ -1,0 +1,19 @@
+import styled from '@emotion/styled';
+
+const StyledContainer = styled.button`
+  display: flex;
+  align-items: center;
+  cursor: pointer;
+  gap: 1.2rem;
+  font-size: 1.6rem;
+`;
+
+const StyledSlackLogo = styled.img`
+  width: 2rem;
+`;
+
+const StyledText = styled.p`
+  font-size: 1.6rem;
+`;
+
+export { StyledContainer, StyledSlackLogo, StyledText };
