+++ frontend/src/layouts/SidebarLayout/components/SidebarSlackMenu/SidebarSlackMenu.tsx
@@ -1,6 +1,6 @@
 import { MouseEventHandler } from 'react';
-import styled from '@emotion/styled';
 import Slack from '../../../../assets/slack.svg';
+import { StyledContainer, StyledSlackLogo, StyledText } from './SidebarSlackMenu.styles';
 
 interface Props {
   onClick: MouseEventHandler<HTMLButtonElement>;
@@ -14,21 +14,5 @@ function SidebarSlackMenu({ onClick }: Props) {
     </StyledContainer>
   );
 }
-
-const StyledContainer = styled.button`
-  display: flex;
-  align-items: center;
-  cursor: pointer;
-  gap: 1.2rem;
-  font-size: 1.6rem;
-`;
-
-const StyledSlackLogo = styled.img`
-  width: 2rem;
-`;
-
-const StyledText = styled.p`
-  font-size: 1.6rem;
-`;
 
 export default SidebarSlackMenu;
