+++ frontend/src/pages/MainPage/components/MainFeatureMenu/MainFeatureMenu.styles.tsx
